refactor: move keystone connect utils
